API Reference

PsiTK.AbstractAlgoInfo — Type

Abstract parent type for all algorithm states for callbacks. Any new algorithm can define a struct that inherits from this. (currently not used)

PsiTK.AdaptiveRandomizedSVD — Type

AdaptiveRandomizedSVD <: ΓCompressionStrategy

Compressing the Coulomb vertex $\Gamma_{mn}^{G}$ through an adaptive randomized SVD.

This algorithm approximates the range of the row space of $\Gamma$ (the orbital indices are considered as superindex) through a thin basis Q, such that

\[\Gamma \approx \Gamma Q Q^\dagger\]

where $\Gamma$ is a $N_{pp} \times N_G$ and $Q$ a $N_G \times N_F$ matrix. This is done through a stochastic Q and a diagonalization of

\[E = -\tilde \Gamma^\dagger \tilde \Gamma = U D U^\dagger\]

where $\tilde \Gamma = \Gamma Q$. The compressed $\Gamma$ is then obtained via $\Gamma_\text{compressed} = \tilde \Gamma U$.

The dimension $N_F$ is found by a preceding adaptive range finder. This finder iteratively increases the columns of Q (i.e. $N_F$) in steps of $\sqrt{N_{pp}}$ and stops when the error for a stochastic test vector $\omega$

\[\varepsilon = \frac{ \Vert (1 - QQ^\dagger)\Gamma^\dagger \omega \Vert}{\Vert \omega \Vert}\]

is smaller than thresh/10.

PsiTK.CoulombGramian — Type

CoulombGramian <: ΓCompressionStrategy

AdaptiveRandomizedSVD

PsiTK.ΓCompressionStrategy — Type

ΓCompressionStrategy

Abstract type for different strategies of compressing the Coulomb vertex $\Gamma_{nm}^G$.

Available models:

CoulombGramian

AdaptiveRandomizedSVD (default)

PsiTK.chain_callbacks — Method

Combines multiple callbacks into one.

PsiTK.compress_coulomb_vertex — Method

compress_coulomb_vertex(ΓmnG; thresh=1e-6,

TODO

PsiTK.compute_coulomb_vertex — Method

Compute the Coulomb vertex

\[Γ_{\bm km, \bm kn \bm G} = \int_Ω \sqrt{\frac{4π}{\bm G^2}} e^{-i\bm r \bm G} \psi_{\bm km}^∗ \psi_{\bm k' n} dr\]

where n_bands is the number of bands to be considered.

PsiTK.dump_cc4s_files — Function

Write CC4S input files from an SCF result scfres or an output from compute_bands to the folder (by default joinpath(pwd(), "cc4s")). This will dump a number of yaml and binary/text data files to this folder. The files written will be returned by the function. If force is true then existing files will be overwritten.

PsiTK.make_lobpcg_callback — Method

Default callback function to dump status of the DFTK.LOBPCG function

Arguments

thresh: the target threshold for the LOBPCG solver

description: a prefix string for the output